Move FlDotPainter from line_chart_data to axis_chart_data
