+++ test/utils/canvas_wrapper_test.dart
@@ -1,6 +1,6 @@
 import 'dart:ui';
 
-import 'package:fl_chart/src/chart/line_chart/line_chart_data.dart';
+import 'package:fl_chart/src/chart/base/axis_chart/axis_chart_data.dart';
 import 'package:fl_chart/src/utils/canvas_wrapper.dart';
 import 'package:fl_chart/src/utils/utils.dart';
 import 'package:flutter_test/flutter_test.dart';
